[jrs] fixed dev setup + fixed some issues with dangling docker/detecting docker being down
